backend/requirements.txt
flask
boto3
face_recognition
insightface
opencv-python
python-dotenv
numpy